backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-multipart
pydantic-settings
pypdf
python-docx
openpyxl
qdrant-client
fastembed
